pyproject.toml
[project]
name = "tui-browser"
version = "0.1.0"
description = "A terminal-based web browser with JavaScript support"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "textual>=0.40",
    "term-image>=0.7",
    "rich>=13.0",
    "Pillow>=10.0",
]

[project.scripts]
tui-browser = "tui_browser.__main__:main"
